Complete Bullshit of Vision
Literally the sample code, copied and pasted into the Vision.java file.
Recommendations: Change the names of all the public voids, delete most
of the goddamn while loops. What we need to find out from here: How to
get an image from the camera that can be processed by the code.